Corrected a couple of bugs and a few enhancements.
Corrected a bug with the start date when zapping the data and file stores.
Corrected a bug where supplying a future day was not caught.
Corrected a bug where a missing year was not caught correctly with the month report.

Added the start and end dates to the title of the All Time Records display.
Made the displaying of command line errors simpler.