Running on python 3.13.7, all libraries up to date.
Two extra columns was added to the data on the 1ast July 2025.
The columns were "Heap" and "Run Time", these were added at the end of the table.

Then two extra columns where added on the 13th of July 2025.
The columns were "VPD" and "10-minute Average Wind Direction", these were added at columns 5 and 20.

So, depending upon the file creation date, different column headers are used when Pandas reads the excel file.
The file creation date is determined when the file is read and stored in the fileStore.

All code passes ruff